rename spack system to spack filesystem
