+++ lib/spack/spack/package.py
@@ -104,7 +104,11 @@ from spack.package_base import (
 from spack.spec import InvalidSpecDetected, Spec
 from spack.util.cpus import determine_number_of_jobs
 from spack.util.executable import *
-from spack.util.system import *
+from spack.util.spack_filesystem import (
+    file_command,
+    mime_type,
+    fix_darwin_install_name
+)
 from spack.variant import (
     any_combination_of,
     auto_or_any_combination_of,
