+++ lib/spack/spack/relocate.py
@@ -25,7 +25,7 @@ import spack.store
 import spack.util.elf as elf
 import spack.util.executable as executable
 import spack.util.path
-import spack.util.system as ssys
+import spack.util.spack_filesystem as ssys
 
 from .relocate_text import BinaryFilePrefixReplacer, TextFilePrefixReplacer
