+++ lib/spack/spack/util/spack_filesystem.py
